# server/src/swagger/routes.yaml
swagger: '3.0.0'
info:
  title: Kids Activity Tracker API
  version: '1.0.0'
  description: |
    Comprehensive API documentation for the Kids Activity Tracker application.
    
    ## API Overview
    
    ### Core Concepts:
    - **Categories**: 5 age-based groupings (Early Years: Parent Participation, Early Years: On My Own, School Age, Youth, All Ages & Family)
    - **Activity Types**: 22 subject-based groupings (Swimming & Aquatics, Team Sports, etc.)
    - **Global Filters**: hideClosedActivities, hideFullActivities (apply to all endpoints except Favorites)
    - **Pagination**: 50 items per page with offset-based pagination
    
    ### Navigation Flows:
    1. **Browse by Location**: Dashboard → Cities → Venues → Activities
    2. **Browse by Category**: Dashboard → Categories → Category Detail → Activities  
    3. **Browse by Activity Type**: Dashboard → Activity Types → Activity Type Detail → Activities
    4. **Recommendations**: Dashboard → Recommended Activities (user-specific)
    
    ### React Native Components Usage:
    - **DashboardScreen**: Uses categories, activity types, locations APIs
    - **CategoryDetailScreen**: Uses `/categories/{id}/activities`
    - **AllCategoriesScreen**: Uses `/categories`
    - **LocationBrowseScreen**: Uses `/locations/cities`, `/locations/{cityId}/venues`
    - **ActivityDetailScreen**: Uses `/activities/{id}`, favorites APIs
    - **SearchScreen**: Uses `/activities/search` with extensive filtering
paths:
...
  /api/child-activities/link:
    post:
      tags: [Child Activities]
      summary: Link child to activity
      description: |
        Create a link between a child and an activity (enrollment, interest, etc.).
        Used by: ActivityDetailScreen, child activity management
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                childId:
                  type: string
                  format: uuid
                activityId:
                  type: string
                  format: uuid
                status:
                  type: string
                  enum: [interested, registered, completed, cancelled]
                notes:
                  type: string
              required: [childId, activityId, status]
      responses:
        201:
          description: Link created successfully
        409:
          description: Link already exists
  /api/child-activities/{childId}/recommendations:
    get:
      tags: [Child Activities]
      summary: Get activity recommendations for child
      description: |
        Returns recommended activities based on child's age, interests, and activity history.
        Used by: RecommendedActivitiesScreen, child-specific recommendations
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: childId
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: hideClosedActivities
          schema:
            type: boolean
            default: true
        - in: query
          name: hideFullActivities
          schema:
            type: boolean
            default: false
      responses:
        200:
          description: Recommended activities for child
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  child:
                    $ref: '#/components/schemas/Child'
                  recommendations:
                    type: array
                    items:
                      type: object
                      properties:
                        activity:
                          $ref: '#/components/schemas/Activity'
                        score:
                          type: number
                          description: Recommendation confidence score
                        reasons:
                          type: array
                          items:
                            type: string
                          description: Why this activity was recommended
  /api/child-activities/bulk-link:
    post:
      tags: [Child Activities]
      summary: Bulk link multiple activities to a child
      description: |
        Create multiple activity links for a child at once.
        Used by: Bulk activity enrollment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                childId:
                  type: string
                  format: uuid
                activities:
                  type: array
                  items:
                    type: object
                    properties:
                      activityId:
                        type: string
                        format: uuid
                      status:
                        type: string
                        enum: [interested, registered, completed, cancelled]
                      notes:
                        type: string
              required: [childId, activities]
      responses:
        201:
          description: Links created successfully
        400:
          description: Validation error
  /api/child-activities/{childId}/activities/{activityId}:
    put:
      tags: [Child Activities]
      summary: Update activity status for child
      description: |
        Update the status of an activity link for a child.
        Used by: Activity management screens
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: childId
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: activityId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [interested, registered, completed, cancelled]
                notes:
                  type: string
      responses:
        200:
          description: Activity link updated
        404:
          description: Link not found
    
    delete:
      tags: [Child Activities]
      summary: Remove activity link from child
      description: |
        Remove the link between a child and an activity.
        Used by: Activity management screens
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: childId
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: activityId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Link removed successfully
        404:
          description: Link not found
  /api/child-activities/history:
    get:
      tags: [Child Activities]
      summary: Get activity history with filters
      description: |
        Get activity history for user's children with filtering options.
        Used by: Activity history screens
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: childId
          schema:
            type: string
            format: uuid
          description: Filter by specific child
        - in: query
          name: status
          schema:
            type: string
            enum: [interested, registered, completed, cancelled]
          description: Filter by activity status
        - in: query
          name: startDate
          schema:
            type: string
            format: date
          description: Filter activities after this date
        - in: query
          name: endDate
          schema:
            type: string
            format: date
          description: Filter activities before this date
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        200:
          description: Activity history
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  history:
                    type: array
                    items:
                      type: object
                  pagination:
                    $ref: '#/components/schemas/PaginationResponse'
  /api/child-activities/{childId}/favorites:
    get:
      tags: [Child Activities]
      summary: Get child's favorite activities
      description: |
        Get activities marked as favorites for a specific child.
        Used by: Child favorites screens
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: childId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Child's favorite activities
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  child:
                    $ref: '#/components/schemas/Child'
                  favorites:
                    type: array
                    items:
                      $ref: '#/components/schemas/Activity'
  /api/child-activities/calendar:
    get:
      tags: [Child Activities]
      summary: Get calendar data for activities
      description: |
        Get activity calendar data with week/month/year views.
        Used by: Calendar screens
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: view
          schema:
            type: string
            enum: [week, month, year]
            default: month
        - in: query
          name: date
          schema:
            type: string
            format: date
          description: Reference date for calendar view
        - in: query
          name: childId
          schema:
            type: string
            format: uuid
          description: Filter by specific child
      responses:
        200:
          description: Calendar data
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  calendar:
                    type: object
  /api/child-activities/stats:
    get:
      tags: [Child Activities]
      summary: Get activity statistics for children
      description: |
        Get comprehensive activity statistics for user's children.
        Used by: Dashboard, statistics screens
      security:
        - bearerAuth: []
      responses:
        200:
          description: Activity statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  stats:
                    type: object
  /api/child-activities/upcoming:
    get:
      tags: [Child Activities]
      summary: Get upcoming activities for notifications
      description: |
        Get upcoming activities for all children for notifications.
        Used by: Notification system, dashboard
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: days
          schema:
            type: integer
            default: 7
          description: Number of days ahead to look
      responses:
        200:
          description: Upcoming activities
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  upcomingActivities:
                    type: array
                    items:
                      type: object
  /api/child-activities/{childId}/activities:
    get:
      tags: [Child Activities]
      summary: Get activities for a specific child
      description: |
        Get all activities associated with a specific child.
        Used by: Child activity screens
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: childId
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: status
          schema:
            type: string
            enum: [interested, registered, completed, cancelled]
          description: Filter by activity status
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        200:
          description: Child's activities
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  child:
                    $ref: '#/components/schemas/Child'
                  activities:
                    type: array
                    items:
                      type: object
                  pagination:
                    $ref: '#/components/schemas/PaginationResponse'
        404:
          description: Child not found

  # ==================== LOCATIONS ENDPOINTS ====================
...
components:
  schemas:
    Activity:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        ageMin:
          type: integer
        ageMax:
          type: integer
        cost:
          type: number
        spotsAvailable:
          type: integer
        totalSpots:
          type: integer
        registrationStatus:
          type: string
          enum: [Open, Closed, Full, Cancelled]
        dateStart:
          type: string
          format: date-time
        dateEnd:
          type: string
          format: date-time
        registrationDate:
          type: string
          format: date-time
        location:
          $ref: '#/components/schemas/Location'
        activityType:
          $ref: '#/components/schemas/ActivityType'
        provider:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
            website:
              type: string
    ActivityType:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        description:
          type: string
        activityCount:
          type: integer
        subtypes:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              name:
                type: string
              code:
                type: string
...
    Location:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        address:
          type: string
        city:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
            province:
              type: string
            country:
              type: string
        activityCount:
          type: integer
...
    Child:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        dateOfBirth:
          type: string
          format: date
        age:
          type: integer
          description: Calculated age in years
        interests:
          type: array
          items:
            type: string
        notes:
          type: string
        isActive:
          type: boolean
...
    PaginationResponse:
      type: object
      properties:
        total:
          type: integer
          description: Total number of items
        limit:
          type: integer
          description: Items per page
        offset:
          type: integer
          description: Items skipped
        pages:
          type: integer
          description: Total number of pages
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...